Fixed warnings

+++ Source/Network/ClientNetwork.cpp
@@ -62,7 +62,7 @@ bool ClientNetwork::Connect()
 	if (!connected)
 		return false;
 	
-	int id = m_packetHandler.StartPack(NetTypeMessageId::ID_PASSWORD_ATTEMPT);
+	uint64_t id = m_packetHandler.StartPack(NetTypeMessageId::ID_PASSWORD_ATTEMPT);
 	m_packetHandler.WriteString(id, m_password.c_str());
 	auto packet = m_packetHandler.EndPack(id);
 	Send(packet);
